Stream pipelines -> The total number of employees https://hyperskill.org/learn/step/2447

+++ practice/src/TheTotalNumberOfEmployees.java
@@ -58,7 +58,7 @@ public class TheTotalNumberOfEmployees {
                 .map(department -> department.getEmployees().stream()
                         .filter(employee -> employee.getSalary() >= threshold)
                         .count())
-                .reduce(0L, (sum, n) -> sum + n).intValue();
+                .reduce(0L, (sum, n) -> sum + n);
     }
 
     // Don't change the code below
